Try prevent upscaling when a too high sizes value is chosen
Chrome and Firefox (but not safari) us the sizes attribute to
set the intrinsic dimensions of the image. If only small images
are available, but the sizes attribute is set very high, the
image will be upscaled. This is butt-ugly.

So we set an upper limit on the sizes attribute of the largest image available.